requirements.txt
Pillow==8.1.2
tk==0.1.0
sv-ttk==0.4.1
